chore: remove fluent api and clean up examples (#254)
