+++ deps/polkadot/types.ts
@@ -1,0 +1,1 @@
+export * from "https://deno.land/x/polkadot@0.0.8/types/mod.ts";
