+++ examples/balance.ts
@@ -4,11 +4,7 @@ import * as U from "../util/mod.ts";
 
 const config = await t.config();
 
-const root = C
-  .chain(config)
-  .pallet("System")
-  .entry("Account", t.alice.publicKey)
-  .read();
+const root = C.readEntry(config, "System", "Account", [t.alice.publicKey]);
 
 console.log(U.throwIfError(await root.run()));
