+++ examples/metadata.ts
@@ -4,10 +4,7 @@ import * as U from "../util/mod.ts";
 
 const config = await t.config();
 
-const root = C
-  .chain(config)
-  .metadata()
-  .read();
+const root = C.metadata(config);
 
 console.log(U.throwIfError(await root.run()));
