--- examples/rpc/subscription.ts
@@ -1,22 +1,0 @@
-import { assert } from "../../deps/std/testing/asserts.ts";
-import { polkadot } from "../../known/mod.ts";
-import * as rpc from "../../rpc/mod.ts";
-
-const client = await rpc.stdClient(polkadot);
-assert(!(client instanceof Error));
-
-let i = 1;
-
-const error = await client.subscribe("chain_subscribeAllHeads", [], (stop) => {
-  return async (message) => {
-    console.log({ [i++]: message.params.result });
-    if (i > 5) {
-      stop();
-      await client.close();
-    }
-  };
-});
-
-if (error) {
-  console.log(error);
-}
