+++ examples/those_of_subxt/read_bonded.ts
@@ -5,6 +5,7 @@ import * as U from "../../util/mod.ts";
 const config = await t.config();
 
 const aliceStash = t.alice.derive("//stash");
+
 const aliceBonded = C.readEntry(config, "Staking", "Bonded", [aliceStash.publicKey]);
 
 console.log(U.throwIfError(await aliceBonded.run()));
