+++ examples/ticker.ts
@@ -1,12 +1,11 @@
-import * as C from "../../mod.ts";
-import * as t from "../../test-util/mod.ts";
-import * as U from "../../util/mod.ts";
+import * as C from "../mod.ts";
+import * as t from "../test-util/mod.ts";
+import * as U from "../util/mod.ts";
 
 const config = await t.config();
 
 const root = C.watchEntry(config, "Timestamp", "Now", [], () => {
   let i = 0;
-
   return (m) => {
     console.log({ [i]: m });
     i++;
