+++ examples/transfer.ts
@@ -2,29 +2,30 @@ import * as C from "../mod.ts";
 import * as t from "../test-util/mod.ts";
 import * as U from "../util/mod.ts";
 
-const config = await t.config();
+const config = await t.config({ altRuntime: "westend" });
 
-const root = C
-  .chain(config)
-  .pallet("Balances")
-  .extrinsic("transfer")
-  .call({
+const root = C.sendAndWatchExtrinsic({
+  config,
+  sender: {
+    type: "Id",
+    value: t.alice.publicKey,
+  },
+  palletName: "Balances",
+  methodName: "transfer",
+  args: {
     value: 12345n,
     dest: {
       type: "Id",
       value: t.bob.publicKey,
     },
-  })
-  .signed({
-    type: "Id",
-    value: t.alice.publicKey,
-  }, (message) => {
+  },
+  sign(message) {
     return {
       type: "Sr25519",
       value: t.alice.sign(message),
     };
-  })
-  .sendAndWatch((stop) => {
+  },
+  createWatchHandler(stop) {
     return (event) => {
       if (typeof event.params.result === "string") {
         console.log("Extrinsic", event.params.result);
@@ -40,7 +41,8 @@ const root = C
         }
       }
     };
-  });
+  },
+});
 
 U.throwIfError(await root.run());
