+++ examples/watch_blocks.ts
@@ -1,14 +1,15 @@
 import * as C from "../mod.ts";
+import * as U from "../util/mod.ts";
 
-console.log(
-  await C.watchBlocks(C.westend, (stop) => {
-    let i = 0;
-    return ({ block }) => {
-      console.log(block.header);
-      if (i === 2) {
-        stop();
-      }
-      i++;
-    };
-  }).run(),
-);
+const root = C.watchBlocks(C.westend, (stop) => {
+  let i = 0;
+  return ({ block }) => {
+    console.log(block.header);
+    if (i === 2) {
+      stop();
+    }
+    i++;
+  };
+});
+
+U.throwIfError(await root.run());
