+++ examples/watch_events.ts
@@ -1,19 +1,15 @@
 import * as C from "../mod.ts";
 import * as U from "../util/mod.ts";
 
-const root = C
-  .chain(C.westend)
-  .pallet("System")
-  .entry("Events")
-  .watch((stop) => {
-    let i = 0;
-    return (event) => {
-      i++;
-      console.log(event);
-      if (i === 5) {
-        stop();
-      }
-    };
-  });
+const root = C.watchEntry(C.rococo, "System", "Events", [], (stop) => {
+  let i = 0;
+  return (event) => {
+    i++;
+    console.log(event);
+    if (i === 5) {
+      stop();
+    }
+  };
+});
 
 U.throwIfError(await root.run());
