--- fluent/mod.ts
@@ -1,10 +1,0 @@
-export * from "./Block.ts";
-export * from "./Call.ts";
-export * from "./Chain.ts";
-export * from "./Entry.ts";
-export * from "./Extrinsic.ts";
-export * from "./Header.ts";
-export * from "./KeyPage.ts";
-export * from "./Metadata.ts";
-export * from "./Pallet.ts";
-export * from "./Signed.ts";
