+++ mod.ts
@@ -2,8 +2,7 @@ export * from "./config/mod.ts";
 export * as $ from "./deps/scale.ts";
 export { BitSequence } from "./deps/scale.ts";
 export * from "./effect/mod.ts";
-export * from "./fluent/mod.ts";
 export * as M from "./frame_metadata/mod.ts";
 export { $era, $null, ChainError, type Era } from "./frame_metadata/mod.ts";
-export { kusama, moonbeam, polkadot, rococo, westend } from "./known/mod.ts";
+export { kusama, moonbeam, polkadot, rococo, westend } from "./known/mod.ts"; // TODO: get rid of this!
 export { type Handle, handle, hex, Iter, mapCreateWatchHandler } from "./util/mod.ts";
